+++ package.json
@@ -33,13 +33,13 @@
     "lucide-react": "^0.474.0",
     "mongodb": "^6.13.0",
     "mongoose": "^8.10.0",
-    "next": "^15.1.5",
+    "next": "14.1.0",
     "next-auth": "^4.24.11",
     "next-images": "^1.8.5",
     "next-themes": "^0.4.4",
     "prisma": "^6.3.1",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-wrap-balancer": "^1.1.1",
     "shadcn-ui": "^0.9.4",
     "sharp": "^0.33.5",
@@ -48,17 +48,17 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
-    "@next/eslint-plugin-next": "^15.1.6",
+    "@next/eslint-plugin-next": "^14.1.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.2.0",
     "@testing-library/user-event": "^14.6.1",
     "@types/jest": "^29.5.14",
-    "@types/react": "19.0.8",
+    "@types/react": "18.2.0",
     "babel-jest": "^29.7.0",
     "chai": "^4.3.7",
     "chai-http": "^4.4.0",
-    "eslint": "^9",
-    "eslint-config-next": "15.1.5",
+    "eslint": "^8.56.0",
+    "eslint-config-next": "14.1.0",
     "eslint-plugin-prettier": "^5.2.3",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
@@ -70,6 +70,6 @@
     "prettier": "^3.4.2",
     "supertest": "^6.3.4",
     "tailwindcss": "^3.4.1",
-    "typescript": "5.7.3"
+    "typescript": "5.3.3"
   }
 }
